fix: conflictos resueltos
