+++ app/src/main/res/layout/activity_conexion_exitosa.xml
@@ -5,7 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/background_gradient"
     tools:context=".ConexionExitosa">
